+++ CHANGELOG.md
@@ -4,9 +4,15 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 About changelog [here](https://keepachangelog.com/en/1.0.0/)
 
+
+## [Unreleased]
+### Changed
+- Automatic test mongod version increased to v7
+
 ## [4.72.3]
 ### Fixed
 - Somatic general case report small variant table can crash with unclassified variants
+
 
 ## [4.72.2]
 ### Changed