+++ CHANGELOG.md
@@ -8,11 +8,13 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 ## [Unreleased]
 ### Changed
 - Automatic test mongod version increased to v7
+### Fixed
+- GnomAD now defaults to hg38 - change build 37 links accordingly
+
 
 ## [4.72.3]
 ### Fixed
 - Somatic general case report small variant table can crash with unclassified variants
-
 
 ## [4.72.2]
 ### Changed